Separate snapshot tests from timetable tests
